pyproject.toml
[project]
name = "stetho-agent"
version = "0.1.0"
description = "AI 기반 건강 가이드 프로토타입 — 폐 청진음 + 생체신호 + 증상 종합 분석"
requires-python = ">=3.11"

dependencies = [
    # UI
    "streamlit>=1.40",
    # LLM
    "langchain>=0.3",
    "langchain-ollama>=0.2",
    "langgraph>=0.2",
    # AI 모델
    "torch>=2.2",
    "transformers>=4.40",
    # 오디오
    "librosa>=0.10",
    "soundfile>=0.12",
    # 시각화
    "plotly>=5.20",
    # 데이터 검증
    "pydantic>=2.0",
    # 설정
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
